LCP PERF: feat. improved contention on matching list
FEAT:
- removed locks in tag list: since task lock is taken, no locking is
  needed within the tag matching data structures.
- hash table to table: tag entries are now read from a table
  preallocated at init time. Again, no lock is needed so mpc hash table
  not needed.

TODO:
- Contention is now located on the task lock. Finer grain locking
  mecanism could be implemented to further improve performance (for
  example a lock on the entries).

+++ src/MPC_Lowcomm/src/lcp/lcp_tag.h
@@ -44,7 +44,7 @@ struct lcp_tag_data {
         int32_t src_tid;
         int32_t dest_tid;
         uint16_t comm;
-        uint64_t seqn;
+        uint16_t seqn;
 };
 //FIXME: move tag header over here.
